fix: add docs for `ccf/tfbd`
fix: use raw output for `art-template`

fix: standardise use of `got`
not using `got()` in one place and `got.get()` in another place
