+++ docs/new-media.md
@@ -4409,18 +4409,6 @@ wechat-feeds 来源[已停止更新](https://github.com/hellodword/wechat-feeds/
 
 </Route>
 
-## 中国计算机学会
-
-### 新闻
-
-<Route author="nczitzk" example="/ccf/news" path="/ccf/news/:category?" :paramsDesc="['分类，见下表，默认为 CCF 新闻']">
-
-| CCF 新闻     | CCF 聚焦 | ACM 信息   |
-| ---------- | ------ | -------- |
-| Media_list | Focus  | ACM_News |
-
-</Route>
-
 ## 中国科学院青年创新促进会
 
 ### 最新博文
