+++ lib/v2/ccf/maintainer.js
@@ -1,0 +1,3 @@
+module.exports = {
+    '/tfbd/:caty/:id': ['tudou027'],
+};
