+++ lib/v2/ccf/radar.js
@@ -1,0 +1,13 @@
+module.exports = {
+    'ccf.org.cn': {
+        _name: '中国计算机学会',
+        tfbd: [
+            {
+                title: '大数据专家委员会',
+                docs: 'https://docs.rsshub.app/study.html#zhong-guo-ji-suan-ji-xue-hui',
+                source: ['/tfbd/:caty/:id', '/'],
+                target: '/ccf/tfbd/:caty/:id',
+            },
+        ],
+    },
+};
