+++ lib/v2/ccf/router.js
@@ -1,0 +1,3 @@
+module.exports = function (router) {
+    router.get('/tfbd/:caty/:id', require('./tfbd/index'));
+};
