+++ lib/v2/ccf/templates/description.art
