+++ lib/v2/ccf/tfbd/index.js
@@ -4,11 +4,11 @@ const cheerio = require('cheerio');
 
 module.exports = async (ctx) => {
     const base = utils.urlBase(ctx.params.caty, ctx.params.id);
-    const res = await got.get(base);
+    const res = await got(base);
     const info = utils.fetchAllArticles(res.data);
     const $ = cheerio.load(res.data);
 
-    const details = await Promise.all(info.map((e) => utils.detailPage(e.link, ctx.cache)));
+    const details = await Promise.all(info.map((e) => utils.detailPage(e, ctx.cache)));
 
     ctx.state.json = {
         info,
@@ -17,10 +17,6 @@ module.exports = async (ctx) => {
     ctx.state.data = {
         title: '大数据专家委员会 - ' + $('.position a:last-child').text(),
         link: base,
-        item: info.map((e, i) => ({
-            title: e.title,
-            description: utils.renderDesc(details[i].description),
-            link: e.link,
-        })),
+        item: details,
     };
 };
