--- lib/v2/tfbd/maintainer.js
@@ -1,3 +1,0 @@
-module.exports = {
-    '/:caty?': ['tudou027'],
-};
