--- lib/v2/tfbd/radar.js
@@ -1,13 +1,0 @@
-module.exports = {
-    'ccf.org.cn': {
-        _name: 'tfbd',
-        tfbd: [
-            {
-                title: '大数据专家委员会',
-                docs: 'https://docs.rsshub.app/government.html#tfbd-da-shu-ju-zhuan-jia-wei-yuan-hui',
-                source: ['/:caty/:id', '/'],
-                target: '/tfbd/:caty/:id',
-            },
-        ],
-    },
-};
