--- lib/v2/tfbd/router.js
@@ -1,3 +1,0 @@
-module.exports = function (router) {
-    router.get('/:caty/:id', require('./index'));
-};
