--- lib/v2/tfbd/templates/description.art
@@ -1,1 +1,0 @@
-{{ desc }}
